refactor: 도메인 예외와 presentation 의 예외를 분리
