+++ kotlin-playground/src/main/kotlin/org/gonza/kotlinplayground/ui/exception/EmptyOrNullableCarNameException.kt
@@ -1,3 +1,3 @@
-package org.gonza.kotlinplayground.domain.car.exception
+package org.gonza.kotlinplayground.ui.exception
 
 class EmptyOrNullableCarNameException : RuntimeException("차량 이름이 존재하지 않습니다.")
