+++ kotlin-playground/src/main/kotlin/org/gonza/kotlinplayground/ui/exception/InvalidTryCountException.kt
@@ -1,3 +1,3 @@
-package org.gonza.kotlinplayground.domain.car.exception
+package org.gonza.kotlinplayground.ui.exception
 
 class InvalidTryCountException : RuntimeException("차량 경주 게임의 횟수가 올바르지 않습니다.")
